fix: align EF Core mapping with singular database schema

+++ backend/src/HorseRacing.Application/Features/ContractAndRegistration/DTOs/RegistrationResponse.cs
@@ -5,7 +5,7 @@ namespace HorseRacing.Application.Features.ContractAndRegistration.DTOs;
 public class RegistrationResponse
 {
     public int Id { get; set; }
-    public int TournamentId { get; set; }
+    public long TournamentId { get; set; }
     public string TournamentName { get; set; } = string.Empty;
     public int HorseId { get; set; }
     public string HorseName { get; set; } = string.Empty;
