+++ backend/src/HorseRacing.Application/Features/ContractAndRegistration/Interfaces/IRegistrationRepository.cs
@@ -7,7 +7,7 @@ namespace HorseRacing.Application.Features.ContractAndRegistration.Interfaces;
 public interface IRegistrationRepository
 {
     Task<Registration?> GetByIdAsync(int id);
-    Task<Registration?> GetByHorseIdAndTournamentIdAsync(int horseId, int tournamentId);
+    Task<Registration?> GetByHorseIdAndTournamentIdAsync(int horseId, long tournamentId);
     Task<IEnumerable<Registration>> GetByOwnerIdAsync(int ownerUserId);
     Task AddAsync(Registration registration);
     Task SaveChangesAsync();
